feat: remove pragyan auth, minor fixes
