+++ src/api/auth/util.rs
@@ -1,4 +1,3 @@
-use crate::api::RedisConn;
 use crate::models::*;
 use crate::schema::user::{self};
 use crate::util::function;
@@ -10,7 +9,6 @@ use jsonwebtoken::{encode, EncodingKey, Header};
 use oauth2::basic::BasicClient;
 use oauth2::{AuthUrl, ClientId, ClientSecret, RedirectUrl, TokenUrl};
 use rand::{distributions::Alphanumeric, Rng};
-use redis::Commands;
 use std::env;
 
 use super::TokenClaims;
@@ -127,56 +125,6 @@ pub fn get_oauth_user(pg_conn: &mut PgConnection, email: &str, name: &str) -> Re
                 function: function!(),
                 error: err,
             })?;
-        Ok(user)
-    }
-}
-
-pub fn get_pragyan_user(
-    pg_conn: &mut PgConnection,
-    redis_conn: &mut RedisConn,
-    email: &str,
-    name: &str,
-) -> Result<User> {
-    // Already logged in before
-    if let Some(user) = user::table
-        .filter(user::email.eq(&email))
-        .first::<User>(pg_conn)
-        .optional()
-        .map_err(|err| DieselError {
-            table: "user",
-            function: function!(),
-            error: err,
-        })?
-    {
-        Ok(user)
-    } else {
-        // First login
-        let random_string: String = rand::thread_rng()
-            .sample_iter(&Alphanumeric)
-            .take(4)
-            .map(char::from)
-            .collect();
-        let username = &format!("{email}_{random_string}");
-        let new_user = NewUser {
-            name,
-            email,
-            username,
-            is_pragyan: &true,
-            attacks_won: &0,
-            defenses_won: &0,
-            trophies: &INITIAL_RATING,
-            avatar_id: &0,
-            artifacts: &0,
-        };
-        let user: User = diesel::insert_into(user::table)
-            .values(&new_user)
-            .get_result(pg_conn)
-            .map_err(|err| DieselError {
-                table: "user",
-                function: function!(),
-                error: err,
-            })?;
-        redis_conn.set(user.id, 0)?;
         Ok(user)
     }
 }
